Split Flow per type state
