+++ src/client/flow/prepare.rs
@@ -1,0 +1,105 @@
+use http::{header, HeaderMap, HeaderName, HeaderValue, Method, Request, Uri, Version};
+
+use crate::body::BodyWriter;
+use crate::client::amended::AmendedRequest;
+use crate::ext::{HeaderIterExt, MethodExt};
+use crate::{ArrayVec, Error};
+
+use super::state::{Prepare, SendRequest};
+use super::{BodyState, CloseReason, Flow, Inner};
+
+impl<B> Flow<B, Prepare> {
+    /// Create a new Flow.
+    pub fn new(request: Request<B>) -> Result<Self, Error> {
+        let mut close_reason = ArrayVec::from_fn(|_| CloseReason::Http10);
+
+        if request.version() == Version::HTTP_10 {
+            // request.analyze() in CallHolder::new() ensures the only versions are HTTP 1.0 and 1.1
+            close_reason.push(CloseReason::Http10)
+        }
+
+        if request.headers().iter().has(header::CONNECTION, "close") {
+            close_reason.push(CloseReason::ClientConnectionClose);
+        }
+
+        let should_send_body = request.method().need_request_body();
+        let await_100_continue = request.headers().iter().has_expect_100();
+
+        let request = AmendedRequest::new(request);
+
+        let default_body_mode = if request.method().need_request_body() {
+            BodyWriter::new_chunked()
+        } else {
+            BodyWriter::new_none()
+        };
+
+        let inner = Inner {
+            request,
+            analyzed: false,
+            state: BodyState {
+                writer: default_body_mode,
+                ..Default::default()
+            },
+            close_reason,
+            should_send_body,
+            await_100_continue,
+            status: None,
+            location: None,
+        };
+
+        Ok(Flow::wrap(inner))
+    }
+
+    /// Inspect call method
+    pub fn method(&self) -> &Method {
+        self.inner.request.method()
+    }
+
+    /// Inspect call URI
+    pub fn uri(&self) -> &Uri {
+        self.inner.request.uri()
+    }
+
+    /// Inspect call HTTP version
+    pub fn version(&self) -> Version {
+        self.inner.request.version()
+    }
+
+    /// Inspect call headers
+    pub fn headers(&self) -> &HeaderMap {
+        self.inner.request.original_request_headers()
+    }
+
+    /// Set whether to allow non-standard HTTP methods.
+    ///
+    /// By default the methods are limited by the HTTP version.
+    pub fn allow_non_standard_methods(&mut self, v: bool) {
+        self.inner.state.allow_non_standard_methods = v;
+    }
+
+    /// Add more headers to the call
+    pub fn header<K, V>(&mut self, key: K, value: V) -> Result<(), Error>
+    where
+        HeaderName: TryFrom<K>,
+        <HeaderName as TryFrom<K>>::Error: Into<http::Error>,
+        HeaderValue: TryFrom<V>,
+        <HeaderValue as TryFrom<V>>::Error: Into<http::Error>,
+    {
+        self.inner.request.set_header(key, value)
+    }
+
+    /// Convert the call to send body despite method.
+    ///
+    /// Methods like GET, HEAD and DELETE should not have a request body.
+    /// Some broken APIs use bodies anyway, and this is an escape hatch to
+    /// interoperate with such services.
+    pub fn send_body_despite_method(&mut self) {
+        self.inner.should_send_body = true;
+        self.inner.state.skip_method_body_check = true;
+    }
+
+    /// Continue to the next flow state.
+    pub fn proceed(self) -> Flow<B, SendRequest> {
+        Flow::wrap(self.inner)
+    }
+}
